chore: remove OpenAPI schema generation references
