+++ package.json
@@ -32,9 +32,7 @@
     "clean": "rm -rf ./.nyc_output ./node_modules/.cache ./coverage",
     "build": "tsup",
     "generate": "tsx gen/generator.ts && pnpm check",
-    "generate-openapi": "tsx gen/generator.ts --source=openapi && pnpm check",
     "generate-local": "tsx gen/generator.ts -- --local && pnpm check",
-    "generate-diff": "tsx gen/generator.ts --source=openapi --output=src/resources-openapi && tsx gen/generator.ts --source=public --output=src/resources-public && diff -r src/resources-openapi src/resources-public && pnpm check",
     "lint": "biome check --enforce-assist=true",
     "check": "biome check --enforce-assist=true --write",
     "semantic-release": "semantic-release",
